Move all Objects to new Laminas\Barcode\Barcode namespace so we keep git history
