+++ src/Barcode/Codabar.php
@@ -6,12 +6,12 @@
  * @license   https://github.com/laminas/laminas-barcode/blob/master/LICENSE.md New BSD License
  */
 
-namespace Laminas\Barcode\Object;
+namespace Laminas\Barcode\Barcode;
 
 /**
  * Class for generate Codabar barcode
  */
-class Codabar extends AbstractObject
+class Codabar extends AbstractBarcode
 {
     /**
      * Coding map
