+++ src/Barcode/Code25.php
@@ -6,12 +6,12 @@
  * @license   https://github.com/laminas/laminas-barcode/blob/master/LICENSE.md New BSD License
  */
 
-namespace Laminas\Barcode\Object;
+namespace Laminas\Barcode\Barcode;
 
 /**
  * Class for generate Interleaved 2 of 5 barcode
  */
-class Code25 extends AbstractObject
+class Code25 extends AbstractBarcode
 {
     /**
      * Coding map
