+++ src/Barcode/Ean8.php
@@ -6,7 +6,7 @@
  * @license   https://github.com/laminas/laminas-barcode/blob/master/LICENSE.md New BSD License
  */
 
-namespace Laminas\Barcode\Object;
+namespace Laminas\Barcode\Barcode;
 
 use Laminas\Validator\Barcode as BarcodeValidator;
